+++ codegen.yml
@@ -48,10 +48,10 @@ generates:
         SemitagResolveEvent: ./SemitagEvent/model.js#SemitagEventModel
         SemitagResolving: ./SemitagResolving/model.js#SemitagResolvingModel
         SemitagSuggestTagsItem: ./SemitagSuggestTagsItem/model.js#SemitagSuggestTagsItemModel
-        SoundcloudVideoSource: ../Soundcloud/dto.js#SoundcloudVideoSourceDTO
-        SoundcloudVideoSourceCreateEvent: ../Soundcloud/dto.js#SoundcloudVideoSourceEventDTO
-        SoundcloudVideoSourceEvent: ../Soundcloud/dto.js#SoundcloudVideoSourceEventDTO
-        SoundcloudVideoSourceEventConnection: ../Soundcloud/dto.js#SoundcloudVideoSourceEventConnectionDTO
+        SoundcloudVideoSource: ../SoundcloudVideoSource/dto.js#SoundcloudVideoSourceDTO
+        SoundcloudVideoSourceCreateEvent: ../SoundcloudVideoSource/dto.js#SoundcloudVideoSourceEventDTO
+        SoundcloudVideoSourceEvent: ../SoundcloudVideoSource/dto.js#SoundcloudVideoSourceEventDTO
+        SoundcloudVideoSourceEventConnection: ../SoundcloudVideoSource/dto.js#SoundcloudVideoSourceEventConnectionDTO
         Tag: ./Tag/model.js#TagModel
         TagEvent: ./TagEvent/model.js#TagEventModel
         TagName: ./TagName/model.js#TagNameModel
