+++ src/SoundcloudVideoSource/SoundcloudVideoSourceEventConnection.resolver.ts
