+++ src/SoundcloudVideoSource/registerMadFromSoundcloud.resolver.test.ts
