+++ src/index.ts
@@ -19,7 +19,7 @@ import { makeResolvers } from "./resolvers/index.js";
 import { CurrentUser, ServerContext, UserContext } from "./resolvers/types.js";
 import { UserModel } from "./resolvers/User/model.js";
 import typeDefs from "./schema.graphql";
-import { SoundcloudService } from "./Soundcloud/service.js";
+import { SoundcloudService } from "./SoundcloudVideoSource/service.js";
 
 const jwksClient = createJwksClient({
   jwksUri: `https://${process.env.AUTH0_DOMAIN}/.well-known/jwks.json`,
